plugins: Remove OpenPGP support
OpenPGP is a poor match for instant messaging. PGP in general is brittle
and complicated to use in practice for end-users. In the early days of
XMPP, it was a good solution, and even when compared with OTR, it had
some benefits. But compared with OMEMO, there is little reason to use
OpenPGP anyway outside of niche use cases better served by other
clients.

Dino aims to be a secure, modern client. To do so, we need to simplify
the UX for end-to-end encryption as much as possible to ensure users do
not misuse encryption and harm themselves in the process. This is a
compelling reason not to support OTR at this point, and it is one reason
that Conversations dropped OTR support (along with code complexity). We
should apply the same treatment to PGP support, and simplify the
codebase in the process.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>